Add remote file opening via HTTP
This adds support for loading the markdown and JSON from remote servers
by using open-uri. This means the docs can live outside of the
application server and not need to be generated or checked into the app.

The usecase for doing this is so that a Continuous Integration (CI)
server can generate the docs since it is already setup for the test
environment, which RAD requires. Then, after generation, it uploads the
RAD docs (JSON and markdown files) to a static file host, like AWS S3,
and then those docs get served to the application environments when
viewing the API docs.

This lets the docs be viewed in live application environments without
checking the docs into the app since they may change often.