Add 'active' classes to active navigation items
We have a somewhat long navigation and it would be super useful to
highlight the active group and the active item in the navigation.

This change allows the styling of active navigation items in CSS
overrides.

The only problem I see currently is that there is already a CSS rule in
the existing Apitome CSS that will make the current example item bold
if this gets merged.

Let me know if you are unhappy with anything and I'm happy to fix any
issues.